moved check to argparse
